src/requirements.txt
astroid==1.3.2
ipython==2.3.1
logilab-common==0.63.2
matplotlib==1.4.2
nose==1.3.4
numpy==1.9.1
pylint==1.4.0
pyparsing==2.0.3
python-dateutil==2.3
pytz==2014.10
pyzmq==14.5.0
scipy==0.14.0
six==1.8.0
